+++ src/contents/sip-page-detail-component.ts
@@ -19,13 +19,13 @@ export class SipPageDetailComponent extends SipComponent {
         let content = `import { Component, ViewContainerRef } from '@angular/core';
 import * as moment from 'moment';
 import { AdChartsModule } from '@delon/abc';
-import { SipPage, SipProvidePage, SipNgInit } from 'sip-alain';
+import { SipPage, SipNgInit } from 'sip-alain';
         
 @Component({
     selector: 'sip-${name}',
     ${template},
     ${style},
-    providers: [...SipProvidePage(${className})]
+    providers: [{ provide: SipBusinessComponent, useExisting: forwardRef(() => ${className}) }]
 })
 export class ${className} extends SipPage {